Move validator functions to validators.py file
